+++ EffectivePotential/src/one_loop_potential.cpp
@@ -130,7 +130,8 @@ Eigen::VectorXd OneLoopPotential::d2V_dxdt(Eigen::VectorXd phi, double T) const
             const auto scalar_masses_sq = get_scalar_masses_sq(phi_shifted, xi);
             const auto fermion_masses_sq = get_fermion_masses_sq(phi_shifted);
             const auto vector_masses_sq = get_vector_masses_sq(phi_shifted);
-
+            const auto ghost_masses_sq = get_ghost_masses_sq(phi_shifted, xi);
+          
             for(int k = 0; k < n_h_xy.size(); ++k)
             {
                 const double T_shifted = T + n_h_xy[k]*h;
@@ -141,7 +142,7 @@ Eigen::VectorXd OneLoopPotential::d2V_dxdt(Eigen::VectorXd phi, double T) const
                 {
                     case DaisyMethod::None:
                     {
-                        V1T_ = V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, T);
+                        V1T_ = V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq, T);
                         gradient(i) += (V1T_ + counter_term_)*coeff_xy[j]*coeff_xy[k] / (h*h);
                         break;
                     }
@@ -150,7 +151,7 @@ Eigen::VectorXd OneLoopPotential::d2V_dxdt(Eigen::VectorXd phi, double T) const
                         const auto scalar_debye_sq = get_scalar_debye_sq(phi_shifted, xi, T_shifted);
                         const auto vector_debye_sq = get_vector_debye_sq(phi_shifted, T_shifted);
                         
-                        V1T_ = V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, T);
+                        V1T_ = V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq, T);
                         daisy_ = daisy(scalar_masses_sq, scalar_debye_sq, vector_masses_sq, vector_debye_sq,
                             T_shifted);
                         gradient(i) += (V1T_ + daisy_ + counter_term_)*coeff_xy[j]*coeff_xy[k] / (h*h);
@@ -161,8 +162,8 @@ Eigen::VectorXd OneLoopPotential::d2V_dxdt(Eigen::VectorXd phi, double T) const
                         const auto scalar_debye_sq = get_scalar_debye_sq(phi_shifted, xi, T_shifted);
                         const auto vector_debye_sq = get_vector_debye_sq(phi_shifted, T_shifted);
 						
-                        V1_= V1(scalar_debye_sq, fermion_masses_sq, vector_debye_sq);
-                        V1T_ = V1T(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, T);
+                        V1_= V1(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, ghost_masses_sq);
+                        V1T_ = V1T(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, ghost_masses_sq, T);
 						
                         gradient(i) += (V1_ + V1T_ + counter_term_)*coeff_xy[j]*coeff_xy[k] / (h*h);
                         break;
@@ -184,30 +185,14 @@ std::vector<double> OneLoopPotential::get_scalar_dofs() const {
 
 double OneLoopPotential::V1(std::vector<double> scalar_masses_sq,
                             std::vector<double> fermion_masses_sq,
-                            std::vector<double> vector_masses_sq) const {
+                            std::vector<double> vector_masses_sq,
+                            std::vector<double> ghost_masses_sq) const {
   double correction = 0;
 
   const auto scalar_dofs = get_scalar_dofs();
   const auto fermion_dofs = get_fermion_dofs();
   const auto vector_dofs = get_vector_dofs();
-  
-  /*if(true)
-	  {
-		  std::cout << "Scalars:";
-		  for(auto i: scalar_masses_sq)
-			  std::cout << " " << i;
-		  std::cout << std::endl;
-		  
-		  std::cout << "Fermions:";
-		  for(auto i: fermion_masses_sq)
-			  std::cout << " " << i;
-		  std::cout << std::endl;
-		  
-		  std::cout << "Bosons:";
-		  for(auto i: vector_masses_sq)
-			  std::cout << " " << i;
-		  std::cout << std::endl;
-	  }*/
+  const auto ghost_dofs = get_ghost_dofs();
 
   if (scalar_dofs.size() != scalar_masses_sq.size()) {
     throw std::runtime_error("Scalar dofs and masses do not match");
@@ -242,15 +227,12 @@ double OneLoopPotential::V1(std::vector<double> scalar_masses_sq,
       (square(renormalization_scale) * xlogx(x) - vector_masses_sq[i] * 5. / 6.);
   }
 
-  // gauge dependent vector correction
-  // hack - i know only first 3 are longitudinal in xSM model
+  // ghost correction
   if (xi != 0.) {
-    for (size_t i = 0; i < vector_masses_sq.size(); ++i) {
-      if (i < 3) {
-        const double x = xi * vector_masses_sq[i] / square(renormalization_scale);
-        correction -= vector_dofs[i] * xi * vector_masses_sq[i] *
-          (square(renormalization_scale) * xlogx(x) - xi * vector_masses_sq[i] * 3. / 2.);
-      }
+    for (size_t i = 0; i < ghost_masses_sq.size(); ++i) {
+        const double x = ghost_masses_sq[i] / square(renormalization_scale);
+        correction -= ghost_dofs[i] * ghost_masses_sq[i] *
+          (square(renormalization_scale) * xlogx(x) - ghost_masses_sq[i] * 3. / 2.);
     }
   }
 
@@ -269,11 +251,13 @@ double OneLoopPotential::V1(Eigen::VectorXd phi, double T) const {
     case DaisyMethod::ArnoldEspinosa:
       return V1(get_scalar_masses_sq(phi, xi),
                 get_fermion_masses_sq(phi),
-                get_vector_masses_sq(phi));
+                get_vector_masses_sq(phi),
+                get_ghost_masses_sq(phi, xi));
     case DaisyMethod::Parwani:
       return V1(get_scalar_debye_sq(phi, xi, T),
                 get_fermion_masses_sq(phi),
-                get_vector_debye_sq(phi, T));
+                get_vector_debye_sq(phi, T),
+                get_ghost_masses_sq(phi, xi));
     default:
       throw std::runtime_error("unknown daisy method");
   }
@@ -281,7 +265,8 @@ double OneLoopPotential::V1(Eigen::VectorXd phi, double T) const {
 
 double OneLoopPotential::V1T(std::vector<double> scalar_masses_sq,
                              std::vector<double> fermion_masses_sq,
-                             std::vector<double> vector_masses_sq, double T) const {
+                             std::vector<double> vector_masses_sq,
+                             std::vector<double> ghost_masses_sq, double T) const {
   double correction = 0;
   
   if(T == 0.)
@@ -290,6 +275,7 @@ double OneLoopPotential::V1T(std::vector<double> scalar_masses_sq,
   const auto scalar_dofs = get_scalar_dofs();
   const auto fermion_dofs = get_fermion_dofs();
   const auto vector_dofs = get_vector_dofs();
+  const auto ghost_dofs = get_ghost_dofs();
 
   if (scalar_dofs.size() != scalar_masses_sq.size()) {
     throw std::runtime_error("Scalar dofs and masses do not match");
@@ -330,13 +316,10 @@ double OneLoopPotential::V1T(std::vector<double> scalar_masses_sq,
   if(T == 0.)
     std::cout << "Correction: " << correction << " after vectors." << std::endl;
 
-  // gauge dependent vector correction
-  // hack - i know only first 3 are longitudinal in xSM model
+  // ghost correction
   if (xi != 0.) {
-    for (size_t i = 0; i < vector_masses_sq.size(); ++i) {
-      if (i < 3) {
-        correction -= vector_dofs[i] * J_B(xi * vector_masses_sq[i] / square(T));
-      }
+    for (size_t i = 0; i < ghost_masses_sq.size(); ++i) {
+        correction -= ghost_dofs[i] * J_B(ghost_masses_sq[i] / square(T));
     }
   }
   
@@ -347,6 +330,7 @@ double OneLoopPotential::V1T(std::vector<double> scalar_masses_sq,
 }
 
 double OneLoopPotential::V1T(Eigen::VectorXd phi, double T) const {
+
   switch (daisy_method) {
     // Arnold-Espinosa method does not alter one-loop finite-temperature potential
     // hence here it is the same as no daisy corrections
@@ -354,11 +338,13 @@ double OneLoopPotential::V1T(Eigen::VectorXd phi, double T) const {
     case DaisyMethod::ArnoldEspinosa:
       return V1T(get_scalar_masses_sq(phi, xi),
                  get_fermion_masses_sq(phi),
-                 get_vector_masses_sq(phi), T);
+                 get_vector_masses_sq(phi),
+                 get_ghost_masses_sq(phi, xi), T);
     case DaisyMethod::Parwani:
       return V1T(get_scalar_debye_sq(phi, xi, T),
                  get_fermion_masses_sq(phi),
-                 get_vector_debye_sq(phi, T), T);
+                 get_vector_debye_sq(phi, T),
+                 get_ghost_masses_sq(phi, xi), T);
     default:
       throw std::runtime_error("unknown daisy method");
   }
@@ -412,32 +398,37 @@ double OneLoopPotential::V(Eigen::VectorXd phi, double T) const {
   const auto scalar_masses_sq = get_scalar_masses_sq(phi, xi);
   const auto fermion_masses_sq = get_fermion_masses_sq(phi);
   const auto vector_masses_sq = get_vector_masses_sq(phi);
-
+  const auto ghost_masses_sq = get_ghost_masses_sq(phi, xi);
+  
   if (T > 0) {
     const auto scalar_debye_sq = get_scalar_debye_sq(phi, xi, T);
     const auto vector_debye_sq = get_vector_debye_sq(phi, T);
     switch (daisy_method) {
       case DaisyMethod::None:
         return V0(phi)
-               + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq)
-               + V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, T)
+               + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq)
+               + V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq, T)
                + counter_term(phi, T);
       case DaisyMethod::ArnoldEspinosa:
         return V0(phi)
                + daisy(scalar_masses_sq, scalar_debye_sq, vector_masses_sq, vector_debye_sq, T)
-               + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq)
-               + V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, T)
+               + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq)
+               + V1T(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq, T)
                + counter_term(phi, T);
       case DaisyMethod::Parwani:
+        if ((scalar_debye_sq.size() != scalar_masses_sq.size())
+             or (vector_debye_sq.size() != vector_debye_sq.size())){
+          throw std::runtime_error("The sizes of scalar_debye_sq and scalar_masses_sq, vector_debye_sq and vector_debye_sq must be equal in Parwani method");
+        }
         return V0(phi)
-               + V1(scalar_debye_sq, fermion_masses_sq, vector_debye_sq)
-               + V1T(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, T)
+               + V1(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, ghost_masses_sq)
+               + V1T(scalar_debye_sq, fermion_masses_sq, vector_debye_sq, ghost_masses_sq, T)
                + counter_term(phi, T);
       default:
         throw std::runtime_error("unknown daisy method");
     }
   } else {
-    return V0(phi) + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq) + counter_term(phi, T);
+    return V0(phi) + V1(scalar_masses_sq, fermion_masses_sq, vector_masses_sq, ghost_masses_sq) + counter_term(phi, T);
   }
 }