+++ example/TestBubbleProfiler/run_BP_2d.cpp
@@ -27,7 +27,9 @@ int main(int argc, char* argv[]) {
   
   // Make ActionCalculator object and calcualte the action
   PhaseTracer::ActionCalculator ac(model);
-  double action = ac.get_action(vacuums[0].x,vacuums[1].x,0); // 0 is the temperature, which is not used in this potential
+  ac.set_action_calculator(PhaseTracer::ActionMethod::BubbleProfiler);
+  
+  double action = ac.get_action(vacuums[0].x,Eigen::VectorXd::Zero(2),0); // 0 is the temperature, which is not used in this potential
   
   std::cout << "S = " << action << std::endl;
   return 0;